Develop to main

+++ TaskPlannerMetrum/Repository/Financs/Finances.cs
@@ -52,6 +52,7 @@ namespace TaskPlannerMetrum.Repository.Financs
                 Guarantee = newfinance.Guarantee,
                 GuaranteePeriod = newfinance.GuaranteePeriod,
                 WorkSpaceID = newfinance.WorkSpaceID,
+               
             });
             _context.SaveChanges();
             int FinanceID = _context.Finances.Where(f => f.ContractID == newfinance.ContractID).OrderBy(i => i.id).Select(f => f.id).LastOrDefault();
@@ -141,7 +142,7 @@ namespace TaskPlannerMetrum.Repository.Financs
                     percents = totalpercents.ToString("0.00").Replace(",", "."),
                     ClientName = _context.Clients.Where(i => i.Id == ClientId).Select(n => n.Name).FirstOrDefault().ToString(),
                     totalbilled = invoicevalues.ToString("N", new System.Globalization.CultureInfo("pt-BR")),
-                    WorkSpace = WorkSpace,
+                 
                     StatusID = blockContract.StatusID,
                     Finances = financas.Select(f => new
                     {
@@ -169,6 +170,8 @@ namespace TaskPlannerMetrum.Repository.Financs
                         f.GuaranteePeriod,
                         f.DateExpectedGarantee,
                         f.StatusGuarantee,
+                        f.WorkspaceName,
+                       
                     })
                 };
                 return contractfinances;
@@ -375,24 +378,44 @@ namespace TaskPlannerMetrum.Repository.Financs
             return _context.Service.Where(i => i.Type == type).ToList();
         }
 
-       
 
 
 
-        public bool DeleteAllService(int ID)
+
+        public bool DeleteAllService(int ID, int userId, string userName)
         {
-           var services = _context.Service.Where(i => i.ID == ID).ToList();        
+            if (string.IsNullOrEmpty(userName))
+            {
+                throw new ArgumentException("O nome do usuário não pode ser nulo ou vazio.");
+            }          
+            var services = _context.Service.Where(i => i.ID == ID).ToList();
+            if (services.Any())
+            {
+                foreach (var service in services)
+                {
+                    var log = new Model.Log
+                    {
+                        UserId = userId,
+                        UserName = userName, 
+                        Action = "DeleteAllService",
+                        Description = $"Serviço deletado: ID {service.ID}, Descrição {service.Description}",
+                        CreatedAt = DateTime.UtcNow,
+                        Status = "Deleted",
+                        SeverityLevel = "Info",
+                        ErrorDetails = null
+                    };                  
+                    _context.Log.Add(log);
+                }
 
-            if (services.Any())            {
-                _context.Service.RemoveRange(services); 
+            
+                _context.Service.RemoveRange(services);
                 _context.SaveChanges();
+
                 return true;
             }
 
-            return false; 
-
+            return false;
         }
-
 
 
         public string UpdateAllService(Service service)
